application/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
matplotlib==3.8.1
numpy==1.26.1
opencv-python==4.8.1.78
Jinja2==3.1.2
python-jose==3.3.0
seaborn==0.13.0
scikit-learn==1.3.2 